Extract pieces from main into new classes
