+++ src/Client/Client.Console/Program.cs
@@ -1,27 +1,16 @@
 ﻿using AnonymousTokensShared;
+using AnonymousTokensShared.Protocol;
 
 using Org.BouncyCastle.Asn1.X9;
 using Org.BouncyCastle.Crypto.Parameters;
-using Org.BouncyCastle.Math;
-using Org.BouncyCastle.Math.EC;
-using Org.BouncyCastle.Security;
 
 using System;
-using System.Collections.Generic;
 using System.Diagnostics;
-using System.Linq;
-using System.Security.Cryptography;
-using System.Text;
-
-using ECCurve = Org.BouncyCastle.Math.EC.ECCurve;
-using ECPoint = Org.BouncyCastle.Math.EC.ECPoint;
 
 namespace AnonymousTokensConsole
 {
     class Program
     {
-        static string ToHex(byte[] data) => string.Concat(data.Select(x => x.ToString("x2")));
-
         /// <summary>
         /// Defines an elliptic curve to be used in our protocol. We will use "secp256k1".
         /// </summary>
@@ -35,230 +24,10 @@ namespace AnonymousTokensConsole
             return ECNamedCurveTable.GetByName(algorithm);
         }
 
-        /// <summary>
-        /// Used by the initiator. Generates an initial point to be submitted to the token service for signing.
-        /// </summary>
-        /// <param name="curve">Curve parameters</param>
-        /// <returns>The seed t for a random point, the initial mask r of the point, and the masked point P</returns>
-        private static (byte[] t, BigInteger r, ECPoint P) Initiate(ECCurve curve)
-        {
-            var random = new SecureRandom();
+        private static readonly Initiator _initiator = new Initiator();
+        private static readonly TokenVerifier _tokenVerifier = new TokenVerifier();
+        private static readonly TokenGenerator _tokenGenerator = new TokenGenerator();
 
-            BigInteger r = RandomCurveNumberGenerator.GenerateRandomNumber(curve, random);
-
-            // Sample random bytes t such that x = hash(t) is a valid
-            // x-coordinate on the curve. Then T = HashToCurve(t).
-            var t = new byte[32];
-            ECPoint T;
-            for (; ; )
-            {
-                random.NextBytes(t);
-                T = HashToCurve(curve, t);
-                if (T == null)
-                    continue;
-                break;
-            }
-
-            // Compute P = r*T
-            ECPoint P = T.Multiply(r);
-            return (t, r, P);
-        }
-
-        /// <summary>
-        /// Used by the token service. Signs the point submitted by the initiator in order to create a token, and outputs a proof of validity.
-        /// </summary>
-        /// <param name="ecParameters">Curve parameters</param>
-        /// <param name="P">Point submitted by the app</param>
-        /// <param name="K">Public key for the token scheme</param>
-        /// <param name="k">Private key for the token scheme</param>
-        /// <returns>A signed point Q and a Chaum-Pedersen proof (c,z) proving that the point is signed correctly</returns>
-        private static (ECPoint Q, BigInteger c, BigInteger z) GenerateToken(X9ECParameters ecParameters, ECPoint P, ECPoint K, BigInteger k)
-        {
-            // Compute Q = k*P
-            var Q = P.Multiply(k);
-
-            // Chaum-Pedersen proof of correct signature
-            var proof = CreateProof(ecParameters, k, K, P, Q);
-
-            return (Q, proof.c, proof.z);
-        }
-
-        /// <summary>
-        /// Hashes a seed t into a point T on the curve. Returns null if t is unsuitable.
-        /// </summary>
-        /// <param name="curve">The elliptic curve in Weierstrass form</param>
-        /// <param name="t">The seed</param>
-        /// <returns>A random point T uniquely determined by seed t, otherwise null</returns>
-        private static ECPoint HashToCurve(ECCurve curve, byte[] t)
-        {
-            ECFieldElement x, ax, x3, y, y2;
-
-            var P = curve.Field.Characteristic;
-            var sha256 = SHA256.Create();
-            var hashAsInt = new BigInteger(sha256.ComputeHash(t));
-
-            // Check that the hash is within valid range
-            if (hashAsInt.CompareTo(BigInteger.One) < 0 || hashAsInt.CompareTo(P) >= 0)
-                return null;
-
-            // A valid point (x,y) must satisfy: y^2 = x^3 + Ax + B mod P
-            // Convert hash from BigInt to FieldElement x modulo P
-            x = curve.FromBigInteger(hashAsInt);    // x
-            ax = x.Multiply(curve.A);               // Ax
-            x3 = x.Multiply(x).Multiply(x);         // x^3
-            y2 = x3.Add(ax).Add(curve.B);           // y^2 = x^3 + Ax + B
-            y = y2.Sqrt();                          // y = sqrt(x^3 + Ax + B)
-
-            // y == null if square root mod P does not exist
-            if (y == null)
-                return null;
-
-            ECPoint T = curve.CreatePoint(x.ToBigInteger(), y.ToBigInteger());
-            return T;
-        }
-
-        /// <summary>
-        /// Used by the initiator. It first verifies that the incoming token is well-formed, and then removes the previously applied mask.
-        /// </summary>
-        /// <param name="ecParameters">Curve parameters</param>
-        /// <param name="K">Public key for the token scheme</param>
-        /// <param name="P">Masked point initially submitted to the token service</param>
-        /// <param name="Q">Signed masked point returned from the token service</param>
-        /// <param name="c">Claimed challenge from the Chaum-Pedersen proof</param>
-        /// <param name="z">Response from the Chaum-Pedersen proof</param>
-        /// <param name="r">Masking of the initial point</param>
-        /// <returns>A randomised signature W on the point chosen by the initiator</returns>
-        private static ECPoint RandomiseToken(X9ECParameters ecParameters, ECPoint K, ECPoint P, ECPoint Q, BigInteger c, BigInteger z, BigInteger r)
-        {
-            // Verify the proof (c,z).
-            if (VerifyProof(ecParameters, K, P, Q, c, z))
-            {
-                Console.WriteLine("Proof is valid.");
-            }
-            else
-            {
-                Console.WriteLine("Proof is not valid.");
-                Debug.Fail("Token is invalid.");
-            }
-
-            // Removing the initial mask r. W = (1/r)*Q = k*T.
-            var rInverse = r.ModInverse(ecParameters.Curve.Order);
-            var W = Q.Multiply(rInverse);
-            return W;
-        }
-
-        /// <summary>
-        /// Used by the token verifier. It recreates the initial point from the initiator, signs it, and verifies that they are equal.
-        /// </summary>
-        /// <param name="curve">Curve parameters</param>
-        /// <param name="t">Seed for the initial point chosen by the initiator</param>
-        /// <param name="W">Token received from the initiator</param>
-        /// <param name="k">Secret key for the token scheme</param>
-        /// <returns>True if the token is valid, otherwise false</returns>
-        private static bool VerifyToken(ECCurve curve, byte[] t, ECPoint W, BigInteger k)
-        {
-            var T = HashToCurve(curve, t);
-            var V = T.Multiply(k);
-            return V.Equals(W);
-        }
-
-        /// <summary>
-        /// Creates the challenge for the Chaum-Pedersen protocol, using the strong Fiat-Shamir transformation.
-	    /// Hashes all input and a fixed domain to create an unpredictable number. Used by the initiator and token service.
-        /// </summary>
-        /// <param name="G">Curve generator</param>
-        /// <param name="P">Randomised point on curve</param>
-        /// <param name="K">Public key K = k*G</param>
-        /// <param name="Q">Signature Q = k*P</param>
-        /// <param name="X">Commitment X = r*G</param>
-        /// <param name="Y">Commitment Y = r*P</param>
-        /// <returns>A random number uniquely based on all inputs</returns>
-        private static BigInteger CreateChallenge(ECPoint G, ECPoint P, ECPoint K, ECPoint Q, ECPoint X, ECPoint Y)
-        {
-            // Encode the ECPoint inputs
-            var GEncoded = G.GetEncoded();
-            var PEncoded = P.GetEncoded();
-            var KEncoded = K.GetEncoded();
-            var QEncoded = Q.GetEncoded();
-            var XEncoded = X.GetEncoded();
-            var YEncoded = Y.GetEncoded();
-
-            // Domain separation: make sure hash is independent of other systems
-            var domain = "smittestopptoken";
-            var domainEncoded = Encoding.ASCII.GetBytes(domain);
-
-            // Using concat() is best for performance: https://stackoverflow.com/a/415396
-            IEnumerable<byte> points = domainEncoded
-                .Concat(GEncoded)
-                .Concat(PEncoded)
-                .Concat(KEncoded)
-                .Concat(QEncoded)
-                .Concat(XEncoded)
-                .Concat(YEncoded);
-
-            var sha256 = SHA256.Create();
-            var hashAsInt = new BigInteger(sha256.ComputeHash(points.ToArray()));
-
-            return hashAsInt.Mod(G.Curve.Order);
-        }
-
-        /// <summary>
-        /// Used by the token service. Creates a full transcript of a Chaum-Pedersen protocol instance, using the strong Fiat-Shamir transform.
-	    /// The Chaum-Pedersen proof proves that the same secret key k is used to compute K = k*G and Q = k*P, without revealing k.
-        /// </summary>
-        /// <param name="ecParameters">Curve parameters</param>
-        /// <param name="k">Secret key for the token scheme, the value of which we prove existence and usage</param>
-        /// <param name="K">Public key for the token scheme</param>
-        /// <param name="P">Point submitted by the initiator</param>
-        /// <param name="Q">Point signed using the secret key</param>
-        /// <returns></returns>
-        private static (BigInteger c, BigInteger z) CreateProof(X9ECParameters ecParameters, BigInteger k, ECPoint K, ECPoint P, ECPoint Q)
-        {
-            var random = new SecureRandom();
-
-            // Sample a random integer 0 < r < N
-            BigInteger r = RandomCurveNumberGenerator.GenerateRandomNumber(ecParameters.Curve, random);
-
-            // Computes X = r*G
-            ECPoint X = ecParameters.G.Multiply(r);
-
-            // Computes Y = r*P
-            ECPoint Y = P.Multiply(r);
-
-            BigInteger c = CreateChallenge(ecParameters.G, P, K, Q, X, Y);
-
-            // Compute proof z = r - ck mod N
-            BigInteger z = r.Subtract(c.Multiply(k)).Mod(ecParameters.Curve.Order);
-
-            return (c, z);
-        }
-
-        /// <summary>
-        /// Used by the initiator. Verifies a transcript of a Chaum-Pedersen protocol instance, using the strong Fiat-Shamir transform.
-        /// </summary>
-        /// <param name="ecParameters">Curve parameters</param>
-        /// <param name="K">Public key for the token scheme</param>
-        /// <param name="P">Point initially submitted by the initiator</param>
-        /// <param name="Q">Point received from the token service</param>
-        /// <param name="c">Claimed challenge from the Chaum-Pedersen proof</param>
-        /// <param name="z">Response from the Chaum-Pedersen proof</param>
-        /// <returns>Returns true if the proof is valid and otherwise returns false</returns>
-        private static bool VerifyProof(X9ECParameters ecParameters, ECPoint K, ECPoint P, ECPoint Q, BigInteger c, BigInteger z)
-        {
-            // Compute X = z*G + c*K = r*G
-            var X = ecParameters.G.Multiply(z).Add(K.Multiply(c));
-
-            // Compute Y = z*P + c*Q = r*P
-            var Y = P.Multiply(z).Add(Q.Multiply(c));
-
-            // Returns true if the challenge from the proof equals the new challenge
-            return c.Equals(CreateChallenge(ecParameters.G, P, K, Q, X, Y));
-        }
-
-        /// Main will be split into three parts:
-        /// Initiator: 		    running Initiate() and RandomiseToken()
-        /// TokenGenerator: 	running GenerateToken()
-        /// TokenVerifier:	    running VerifyToken()
         static void Main(string[] args)
         {
             // Import parameters for the elliptic curve secp256k1
@@ -270,23 +39,23 @@ namespace AnonymousTokensConsole
             var publicKey = keyPair.Public as ECPublicKeyParameters;
 
             // Initiate communication with a masked point P = r*T = r*Hash(t)
-            var init = Initiate(ecParameters.Curve);
+            var init = _initiator.Initiate(ecParameters.Curve);
             var t = init.t;
             var r = init.r;
             var P = init.P;
 
             // Generate token Q = k*P and proof (c,z) of correctness
-            var token = GenerateToken(ecParameters, P, publicKey.Q, privateKey.D);
+            var token = _tokenGenerator.GenerateToken(ecParameters, P, publicKey.Q, privateKey.D);
             var Q = token.Q;
             var c = token.c;
             var z = token.z;
 
             // Randomise the token Q, by removing the mask r: W = (1/r)*Q = k*T.
             // Also checks that proof (c,z) is correct.
-            var W = RandomiseToken(ecParameters, publicKey.Q, P, Q, c, z, r);
+            var W = _initiator.RandomiseToken(ecParameters, publicKey.Q, P, Q, c, z, r);
 
             // Verify that the token (t,W) is correct.
-            if (VerifyToken(ecParameters.Curve, t, W, privateKey.D))
+            if (_tokenVerifier.VerifyToken(ecParameters.Curve, t, W, privateKey.D))
             {
                 Console.WriteLine("Token is valid.");
             }
